Push de toute la libft finit et corrigee

+++ ft_strsub.c
@@ -6,7 +6,7 @@
 /*   By: hponcet <marvin@42.fr>                     +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2015/12/08 11:46:13 by hponcet           #+#    #+#             */
-/*   Updated: 2015/12/15 23:24:01 by hponcet          ###   ########.fr       */
+/*   Updated: 2015/12/17 21:25:30 by hponcet          ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -19,7 +19,7 @@ char		*ft_strsub(char const *s, unsigned int start, size_t len)
 	size_t		j;
 
 	j = 0;
-	buf = ft_strnew((unsigned int)len);
+	buf = ft_strnew(len);
 	i = (size_t)start + len;
 	if (s == NULL || buf == NULL)
 		return (NULL);
